Fix feedbacks from PR review
