+++ src/main/environment/common_dev.properties
@@ -18,7 +18,7 @@ carestreamOrderCreateURL =http://10.208.122.32:8080/hwc-commonapi-v1.0/carestrea
 registrationUrl =http://10.208.122.32:8080/hwc-commonapi-v1.0/beneficiary/create
 
 #syncSearchByLocation =http://localhost:8082/id/searchByDistrictId
-syncSearchByLocation =http://10.208.122.32:8082/id/searchByDistrictId
+syncSearchByLocation =http://10.208.122.32:8080/hwc-identity-0.0.1/id/searchByDistrictId
 
 #registrarQuickSearchByIdUrl =http://localhost:8080/beneficiary/searchUserByID
 registrarQuickSearchByIdUrl =http://10.208.122.32:8080/hwc-commonapi-v1.0/beneficiary/searchUserByID
