+++ src/main/environment/common_local.properties
@@ -15,7 +15,7 @@ carestreamOrderCreateURL =http://10.208.122.38:8080/commonapi-v1.0/carestream/cr
 registrationUrl =http://10.208.122.38:8080/commonapi-v1.0/beneficiary/create
 
 #syncSearchByLocation =http://localhost:8082/id/searchByDistrictId
-syncSearchByLocation =http://10.208.122.32:8082/id/searchByDistrictId
+syncSearchByLocation =http://10.208.122.38:8080/identity-0.0.1/id/searchByDistrictId
 
 #registrarQuickSearchByIdUrl =http://localhost:8080/beneficiary/searchUserByID
 registrarQuickSearchByIdUrl =http://10.208.122.38:8080/commonapi-v1.0/beneficiary/searchUserByID
