+++ src/main/java/com/iemr/hwc/fhir/dto/beneficiary/identityDTO/BenDetailDTO.java
@@ -19,7 +19,7 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see https://www.gnu.org/licenses/.
 */
-package com.iemr.hwc.fhir.dto.beneficiary;
+package com.iemr.hwc.fhir.dto.beneficiary.identityDTO;
 
 import lombok.Data;
